+++ CHANGELOG.md
@@ -68,6 +68,8 @@ and this project adheres to
   - [#2443](https://github.com/iovisor/bpftrace/pull/2443)
 - Fix BTF detection macro in tools/old/mdflush.bt
   - [#2444](https://github.com/iovisor/bpftrace/pull/2444)
+- Clang parser: resolve unknown enum values from BTF
+  - [#2472](https://github.com/iovisor/bpftrace/pull/2472)
 
 #### Docs
 #### Tools